Error handling, routes and state management added

+++ Frontend/src/components/GigDetails.js
@@ -1,13 +1,21 @@
 import { useGigsContext } from "../hooks/useGigsContext"
+import { useAuthContext } from "../hooks/useAuthContext"
 
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const GigDetails = ({ gig }) => {
     const { dispatch } = useGigsContext()
+    const { user } = useAuthContext()
 
     const handleClick = async () => {
+        if (!user) {
+            return
+        }
         const response = await fetch('/api/gigs/' + gig._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
